# [iOS 17 perf degradation fix] Changing sanitizingNamespace implementation to use regular expression pattern matching.