@rafael-assis 2023-10-11
> Many thanks @rafael-assis . Merged

Great! Thank you so much for merging it so quickly @nalexn! 🙏 
We'll keep experimenting in our codebase and let you know if we can find other perf issues (and fixes for them).